Enable -Werror=sign-compare if available
And fix violations in tests and benchmarks, particularly:
* Use size_t when ranging over sizes
* Use uint64_t in tests where the underlying apis do, as in the case of
  ConsumeDecimalNumber and GetFileSize
* Make ConcurrentTest::K an int because there is no use that calls for
  it to be `uint32_t`

+++ util/cache_test.cc
@@ -158,11 +158,11 @@ TEST_F(CacheTest, UseExceedsCacheSize) {
   }
 
   // Check that all the entries can be found in the cache.
-  for (int i = 0; i < h.size(); i++) {
+  for (size_t i = 0; i < h.size(); i++) {
     ASSERT_EQ(2000 + i, Lookup(1000 + i));
   }
 
-  for (int i = 0; i < h.size(); i++) {
+  for (size_t i = 0; i < h.size(); i++) {
     cache_->Release(h[i]);
   }
 }
